requirements.txt
fastapi==0.115.11
pydantic[email]==2.10.6
pymysql==1.1.1
uvicorn==0.30.6
SQLAlchemy==2.0.39
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt==2.9.0
cryptography==44.0.2
pycryptodome==3.21.0


